Add `HttpClient` abstract class
