+++ src/test/java/com/stripe/net/HttpURLConnectionClientTest.java
@@ -10,7 +10,7 @@ import com.stripe.net.RequestOptions.RequestOptionsBuilder;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
-public class HttpClientTest {
+public class HttpURLConnectionClientTest {
   @Test
   public void testAppInfo() {
     final RequestOptions options = new RequestOptionsBuilder().setApiKey("sk_foobar").build();
@@ -18,7 +18,7 @@ public class HttpClientTest {
     Stripe.setAppInfo(
         "MyAwesomePlugin", "1.2.34", "https://myawesomeplugin.info", "pp_partner_1234");
 
-    final Map<String, String> headers = HttpClient.getHeaders(options);
+    final Map<String, String> headers = HttpURLConnectionClient.getHeaders(options);
 
     final String expectedUserAgent =
         String.format(
@@ -46,9 +46,12 @@ public class HttpClientTest {
   public void testStripeVersion() {
     final RequestOptions versionOverrideOpts =
         new RequestOptionsBuilder().setStripeVersionOverride("2015-05-05").build();
-    assertEquals("2015-05-05", HttpClient.getHeaders(versionOverrideOpts).get("Stripe-Version"));
+    assertEquals(
+        "2015-05-05",
+        HttpURLConnectionClient.getHeaders(versionOverrideOpts).get("Stripe-Version"));
 
     final RequestOptions options = new RequestOptionsBuilder().build();
-    assertEquals(Stripe.API_VERSION, HttpClient.getHeaders(options).get("Stripe-Version"));
+    assertEquals(
+        Stripe.API_VERSION, HttpURLConnectionClient.getHeaders(options).get("Stripe-Version"));
   }
 }
